+++ examples/binding_test.rs
@@ -78,7 +78,7 @@ fn main() {
     window.update(&sim_display);
     let window_rc = Rc::new(RefCell::new(window));
 
-    let mut display = Display::new(HOR_RES as i32, VER_RES as i32);
+    let mut display = Display::new(HOR_RES, VER_RES);
 
     let buffer = DrawBuffer::<{ HOR_RES * LINE_HEIGHT }, Rgb565>::new(HOR_RES, LINE_HEIGHT);
