+++ examples/no_ecs.rs
@@ -67,7 +67,7 @@ fn main() {
     info!("SIMULATOR OK");
     error!("Random error");
 
-    let mut display = Display::new(HOR_RES as i32, VER_RES as i32);
+    let mut display = Display::new(HOR_RES, VER_RES);
 
     let buffer = DrawBuffer::<{ HOR_RES * LINE_HEIGHT }, Rgb565>::new(HOR_RES, LINE_HEIGHT);
